Split misc.dialogs into separate modules
This commit breaks module misc.dialogs into one submodule for each dialog. The module
itself becomes a package re-exporting the classes so the interface remains the same.

Closes: #96
