+++ hamster_gtk/misc/dialogs/__init__.py
@@ -15,9 +15,11 @@
 # You should have received a copy of the GNU General Public License
 # along with 'hamster-gtk'.  If not, see <http://www.gnu.org/licenses/>.
 
-"""This module provides several multi purpose dialoges and widgets."""
+"""This module provides several multi purpose dialogs."""
 
 from __future__ import absolute_import, unicode_literals
 
-from .dialogs import (DateRangeSelectDialog, EditFactDialog,  # NOQA
-                      ErrorDialog, HamsterAboutDialog)
+from .hamster_about_dialog import HamsterAboutDialog  # NOQA
+from .date_range_select_dialog import DateRangeSelectDialog  # NOQA
+from .edit_fact_dialog import EditFactDialog  # NOQA
+from .error_dialog import ErrorDialog  # NOQA
