+++ hamster_gtk/misc/dialogs/error_dialog.py
@@ -15,9 +15,26 @@
 # You should have received a copy of the GNU General Public License
 # along with 'hamster-gtk'.  If not, see <http://www.gnu.org/licenses/>.
 
-"""This module provides several multi purpose dialoges and widgets."""
+
+"""This module contains Dialog for displaying an error message."""
+
 
 from __future__ import absolute_import, unicode_literals
 
-from .dialogs import (DateRangeSelectDialog, EditFactDialog,  # NOQA
-                      ErrorDialog, HamsterAboutDialog)
+from gi.repository import Gtk
+
+
+class ErrorDialog(Gtk.MessageDialog):
+    """
+    Error Dialog used to provide feedback to the user.
+
+    Right now we basicly just pass exception messages to the user. Whilst this is usefull to
+    provide meaningfull information at this early stage, we probably want to provide more user
+    friendly messages later on.
+    """
+
+    def __init__(self, parent, message, *args, **kwargs):
+        """Initialize dialog."""
+        super(ErrorDialog, self).__init__(*args, buttons=Gtk.ButtonsType.CLOSE,
+            message_type=Gtk.MessageType.ERROR, text=message, **kwargs)
+        self.set_transient_for(parent)
